+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ### Added
 - `fx::freeze_at`: freezes another effect at a specific alpha (transition) value.
+- `fx::remap_alpha`: rescales an effect's alpha progression to a smaller window.
 - trait `IntoTemporaryEffect`: previously not exposed although it was implemented for `Effect`.
 
 ### DSL
+- the parsers now recognize boolean values.
 - `.with_duration()` is now available on effects.
 - `fx::freeze_at` added to the DSL.
+- `fx::remap_alpha` added to the DSL.
 
 ### Changed`
 - `fx::explode`: cells "behind" the explosion now have `Color::Black` instead of `Color::Reset` for
@@ -20,6 +23,7 @@
   - Filters set on individual effects won't be overwritten during effect composition
   - Custom `Shader` implementations will need to update to store `cell_filter` as `Option<CellFilter>`
   - `default_shader_impl!(@filter)` macro users must update their field type from `CellFilter` to `Option<CellFilter>`
+
 
 ## tachyonfx 0.14.0 - 2025-04-21